src/main/java/com/bestshawn/basic/BinarySearch.java
package com.bestshawn.basic;

public class BinarySearch {
    /**
     * 二分法查找
     * @param sortedArr
     * @param num
     * @return
     */
    public boolean exists(int[] sortedArr, int num) {
        if (sortedArr == null || sortedArr.length == 0) {
            return false;
        }
        int left = 0;
        int right = sortedArr.length - 1;
        int mid = 0;
        while (left <= right) {
            mid = left + ((right - left) >> 1);
            if (sortedArr[mid] == num) {
                return true;
            } else if (sortedArr[mid] > num) {
                right = mid - 1;
            } else {
                left = mid + 1;
            }
        }

        return false;
    }
}
